feat: revoke token when deleting user
